+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Cover more functional tests on back end home page and left menu
 - Add a Protractor "coreCommon" suite to tests generic features common to plugins and core. Each plugin should execute this suite with their functional tests
 - Cover more functional tests on back end profile page
+- Cover more functional tests on back end Web Service page
 
 ## DEPENDENCIES